refactor proofs stores (#257)
* refactor proofs stores

* update ProofStore

* fix input
